TS-111, TS-55: TYPO3.Neos:Page: custom js/css now appended independently from site package
This eliminates the need to include these styles/js in the site package - they are now included always, on the TYPO3.Neos:Page level.

Additionally, they are now included at the *end* of BODY, not in HEAD section. This resolves the problem when the styles were not applied without complete reload of the backend.